Add files via upload
This definitely is not finished, but it is a little more refined. If you can see any obvious syntax errors let me know, but please don't feel like I want you to actually debug it. I'm just trying to make sure the structure I'm building on isn't totally fucked before building up the file to pull in the data.
The comments should give you a good indication of where I'm having a hard time. There's a specific one on SMA10 and EMA12 about the structure of the .append(). If you could give any advice on which one would be most efficient, or if one version just flat won't work, I'd greatly appreciate it.